+++ src/Entity/Wms_inventory.cs
@@ -90,7 +90,7 @@ namespace YL.Core.Entity
         /// Nullable:False
         /// </summary>
         [SugarColumn()]
-        public byte IsDel { get; set; }
+        public int IsDel { get; set; }
 
         /// <summary>
         /// Desc:创建人
